refactoring AST module & add/update unit tests (move all unit tests to tests dir)
