+++ src/ast/environment_variable_data.rs
@@ -25,22 +25,3 @@ impl TryFrom<Pair<'_, Rule>> for EnvironmentVariableData {
         })
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::test_helpers::*;
-
-    #[test]
-    fn envvar_data_test() {
-        let def = "
-ENVVAR_DATA_ SomeEnvVarData: 399;
-";
-        let exp = EnvironmentVariableData {
-            env_var_name: "SomeEnvVarData".to_string(),
-            data_size: 399,
-        };
-        let val = test_into::<EnvironmentVariableData>(def.trim_start(), Rule::env_var_data);
-        assert_eq!(val, exp);
-    }
-}
