+++ src/ast/multiplex_indicator.rs
@@ -1,5 +1,3 @@
-use std::str;
-
 use crate::parser::DbcError;
 
 #[derive(Copy, Clone, Debug, PartialEq, Eq, Hash)]
@@ -38,26 +36,5 @@ impl TryFrom<&str> for MultiplexIndicator {
         }
 
         Err(Self::Error::UnknownMultiplexIndicator(text.to_string()))
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn multiplexer_indicator_test() {
-        let val: MultiplexIndicator = "m34920".try_into().unwrap();
-        assert_eq!(val, MultiplexIndicator::MultiplexedSignal(34920));
-
-        let val: MultiplexIndicator = "M".try_into().unwrap();
-        assert_eq!(val, MultiplexIndicator::Multiplexor);
-
-        // Empty string is not a valid multiplexer indicator, so we skip this test
-        // let val: MultiplexIndicator = "".try_into().unwrap();
-        // assert_eq!(val, MultiplexIndicator::Plain);
-
-        let val: MultiplexIndicator = "m8M".try_into().unwrap();
-        assert_eq!(val, MultiplexIndicator::MultiplexorAndMultiplexedSignal(8));
     }
 }
