+++ src/ast/version.rs
@@ -17,20 +17,3 @@ impl TryFrom<Pair<'_, Rule>> for Version {
         Ok(Self(inner_str(v)))
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::test_helpers::*;
-
-    #[test]
-    fn version_test() {
-        let def = r#"
-VERSION "HNPBNNNYNNNNNNNNNNNNNNNNNNNNNNNNYNYYYYYYYY>4>%%%/4>'%**4YYY///"
-"#;
-        let exp =
-            Version("HNPBNNNYNNNNNNNNNNNNNNNNNNNNNNNNYNYYYYYYYY>4>%%%/4>'%**4YYY///".to_string());
-        let val = test_into::<Version>(def.trim_start(), Rule::version);
-        assert_eq!(val, exp);
-    }
-}
