+++ tests/unit/main.rs
@@ -1,0 +1,28 @@
+//! AST and pest-rule unit tests (formerly `#[cfg(test)]` modules in `src/`).
+
+mod common;
+
+mod attribute_default;
+mod attribute_definition;
+mod attribute_value_for_object;
+mod byte_order;
+mod comment;
+mod environment_variable;
+mod environment_variable_data;
+mod extended_multiplex;
+mod lexer;
+mod message;
+mod message_id;
+mod message_transmitter;
+mod multiplex_indicator;
+mod node;
+mod numeric_value;
+mod signal;
+mod signal_extended_value_type_list;
+mod signal_groups;
+mod symbol;
+mod val_description;
+mod value_description;
+mod value_table;
+mod value_type;
+mod version;
